+++ sematic/examples/retry/__main__.py
@@ -1,0 +1,5 @@
+# Sematic
+from sematic.examples.retry.pipeline import raise_exception
+
+if __name__ == "__main__":
+    raise_exception().resolve()
